app/requirements.txt
streamlit==1.39.0
databricks-sdk==0.40.0
databricks-sql-connector==3.7.3
pandas==2.2.3
plotly==5.24.1
